# Optimize Huber loss by simpler calculation