--- a/chainer/functions/loss/huber_loss.py
+++ b/chainer/functions/loss/huber_loss.py
@@ -31,10 +31,16 @@ class HuberLoss(function_node.FunctionNode):
         xp = cuda.get_array_module(*inputs)
         x0, x1 = inputs
         diff = x0 - x1
+        delta = diff.dtype.type(self.delta)
+
+        xp.abs(diff, out=diff)
         y = xp.square(diff)
-        mask = y > (self.delta ** 2)
-        y -= mask * xp.square(abs(diff) - self.delta)
+        diff -= delta
+        xp.maximum(diff, 0, dtype=diff.dtype, out=diff)
+        xp.square(diff, out=diff)
+        y -= diff
         y *= 0.5
+
         if self.reduce == 'sum_along_second_axis':
             return y.sum(axis=1),
         else:
@@ -42,14 +48,12 @@ class HuberLoss(function_node.FunctionNode):
 
     def backward(self, indexes, grad_outputs):
         x0, x1 = self.get_retained_inputs()
-        diff = x0 - x1
         gy, = grad_outputs
+        diff = x0 - x1
+        # `functions.clip` only accepts float value.
+        delta = float(self.delta)
 
-        xp = cuda.get_array_module(diff)
-        mask = xp.abs(diff.array) <= self.delta
-
-        gx = chainer.functions.where(
-            mask, diff, self.delta * xp.sign(diff.array))
+        gx = chainer.functions.clip(diff, -delta, delta)
 
         if self.reduce == 'sum_along_second_axis':
             gy = gy.reshape(gy.shape + (1,) * (diff.ndim - 1))