Forward and backward computations of Huber loss function are optimized.

* `xp.abs` is used instead of `abs`,
* `out` arguments are aggressively used for efficient memory access,
* `chainer.function.clip` is used for the backward computation.

CUDA kernel would achieve better performance, but it is out of scope of this PR.