@niboshi 2018-05-15
Thank you for the improvement!

In my benchmark, backward is faster but forward is not.
Could you check forward computation?

master
```
forward             :    55.284 us   +/- 6.720 (min:   51.980) us
backward            :   248.941 us   +/- 3.165 (min:  245.336) us         

forward             :    56.646 us   +/- 6.633 (min:   53.218) us
backward            :   254.573 us   +/-20.623 (min:  246.579) us        

forward             :    55.700 us   +/- 7.059 (min:   51.902) us
backward            :   320.981 us   +/-35.906 (min:  251.722) us
```                                                                                            

This PR
```
forward             :    59.320 us   +/- 5.454 (min:   57.041) us
backward            :   235.110 us   +/- 4.020 (min:  231.728) us

forward             :    57.420 us   +/- 4.797 (min:   55.432) us
backward            :   234.781 us   +/-21.002 (min:  226.693) us

forward             :    59.093 us   +/- 5.345 (min:   56.486) us
backward            :   238.777 us   +/- 4.667 (min:  235.268) us
```