@grafi-tt 2018-05-17
@niboshi Thank you for your performance evaluation! I conducted a benchmark on my Linux machine, but this PR consistently perform better than the master branch both on CPU and GPU. Could you share your benchmark?

Tested software versions are Python 3.6, CuPy 4.0.0 and CUDA 9.0. The machine has Core i5-3550 and GTX 1080Ti. My (arguably sloppy) benchmark script is as below.

```
import chainer
import cupy
import numpy
import time


def bench_gpu():
    x0 = cupy.arange(1024 * 1024, dtype=cupy.float32).reshape(1024, 1024)
    x1 = cupy.arange(1024 * 1024, dtype=cupy.float32).reshape(1024, 1024)
    time.sleep(1)

    start = cupy.cuda.Event()
    end = cupy.cuda.Event()
    start.record()
    l = chainer.functions.huber_loss(x0, x1, delta=1.)
    end.record()
    end.synchronize()
    t = cupy.cuda.get_elapsed_time(start, end)
    print(t)


def bench_cpu():
    x0 = numpy.arange(1024 * 1024, dtype=numpy.float32).reshape(1024, 1024)
    x1 = numpy.arange(1024 * 1024, dtype=numpy.float32).reshape(1024, 1024)

    start = time.perf_counter()
    l = chainer.functions.huber_loss(x0, x1, delta=1.)
    end = time.perf_counter()
    t = end - start
    print(t)


with chainer.no_backprop_mode():
    print("gpu")
    for i in range(10):
        bench_gpu()
    print("cpu")
    for i in range(10):
        bench_cpu()
```