@niboshi 2018-05-17
Ah, I'm sorry, I was benchmarking with very small arrays.
Now I tested with large arrays as in your benchmark, and confirmed both forward and backward computation are improved.

master
```
forward             :  8629.765 us   +/-784.995 (min: 8297.305) us
backward            : 12835.636 us   +/-1123.289 (min:12338.551) us
```
This PR (original)
```
forward             :  7343.509 us   +/-729.469 (min: 6985.093) us
backward            :  5949.416 us   +/-588.071 (min: 5639.520) us
```
This PR (https://github.com/chainer/chainer/pull/4775#issuecomment-389757286)
```
forward             :  7365.982 us   +/-662.457 (min: 7063.958) us
backward            :  5873.546 us   +/-667.226 (min: 5477.769) us
```
(I repeated measurements, and it seems this PR (original) is slightly but consistently faster than https://github.com/chainer/chainer/pull/4775#issuecomment-389757286).